refactor: keep trace public preview audit clean
